PEP fixes
